requirements.txt
python-telegram-bot>=20.0
openai-whisper>=20240314
pyyaml>=6.0
